fix(alerts): use INT UNSIGNED PKs to match Pterodactyl convention
Mixing bigIncrements() (BIGINT UNSIGNED) on panel_alerts with the FKs
into existing Pterodactyl tables (users.id, nodes.id — both INT
UNSIGNED) tripped MySQL errno 150 when adding the
panel_alert_dismissals.alert_id → panel_alerts.id constraint. The
exact root cause is environment-specific (probably a default-charset
clash on the BIGINT column), but switching to increments() /
unsignedInteger() everywhere matches the rest of the schema and side-
steps the issue cleanly.

Also split the over-wide composite index on panel_alerts into two
narrower indexes (scope+node_id, starts_at+ends_at).

+++ database/migrations/2026_04_25_000010_create_panel_alert_dismissals_table.php
@@ -8,8 +8,8 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::create('panel_alert_dismissals', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('alert_id');
+            $table->increments('id');
+            $table->unsignedInteger('alert_id');
             $table->unsignedInteger('user_id');
             $table->timestamp('dismissed_at')->useCurrent();
